refactor: DRY validator middleware
